refactor: adjust id type to Long

+++ src/main/java/br/com/alelo/consumer/consumerpat/respository/ExtractRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.alelo.consumer.consumerpat.domain.entity.Extract;
 
-public interface ExtractRepository extends JpaRepository<Extract, Integer> {
+public interface ExtractRepository extends JpaRepository<Extract, Long> {
 }
